Use `alist-get` to converting alist to hash table
Using `seq-map` to iterate over an alist an inserting the car-cdr-pair
as each came ends up calling `puthash` for each cons in order; however,
the idiomatic way to override previous entries in an alist is to insert
a new entry for a key earlier in the list. Calling `puthash` in order
causes the opposite behavior and ends up using the last value for each
key instead of the first.

This uses `alist-get` to grab the value for each key instead of assuming
that the value should be the same as the current cdr, so that the
resulting hash table key-value mappings as the original alist.